+++ Facturacion.Domain/DTOs/ProductUpdateDTO.cs
@@ -21,8 +21,8 @@ namespace Facturacion.Domain.DTOs
         [Range(1, double.MaxValue, ErrorMessage = "El valor de {0} debe ser mayor a 0")]
         public decimal Price { get; set; }
         [Required]
-        [Range(1, int.MaxValue, ErrorMessage = " {0} debe de ser mayor a 0}")]
-        public int IdCategory { get; set; }
+        [Range(1, double.MaxValue, ErrorMessage = " {0} debe de ser mayor a 0}")]
+        public double IdCategory { get; set; }
         [Required]
         [Range(1, int.MaxValue, ErrorMessage = " {0} debe de ser mayor a 0}")]
         public int Stock { get; set; }
